TheLake/requirements.txt
aniso8601==8.1.0
appdirs==1.4.3
APScheduler==3.6.3
Brotli==1.0.9
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-API==2.0
Flask-Compress==1.8.0
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.18.2
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mongoengine==0.22.1
msgpack==0.6.2
numpy==1.19.5
packaging==20.3
pandas==1.2.0
passlib==1.7.4
patsy==0.5.1
pep517==0.8.2
plotly==4.14.3
plotly-express==0.4.1
progress==1.5
psycopg2-binary==2.8.6
py4j==0.10.9
PyJWT==1.7.1
pymongo==3.11.2
pyparsing==2.4.6
pyspark==3.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.5
pywebhdfs==0.4.1
requests==2.25.1
retrying==1.3.3
scipy==1.6.0
six==1.15.0
statsmodels==0.12.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
